Created first version of models
This includes classes for reports, users, stars (or upvotes). Some
functions were tested with some basic HTML output.  A lot more testing
needs to be done and a lot of other classes and properties still need
to be added - for example, locations for the reports, a password for
users (and validation for everything), an investigation class, etc.